Added Classifier to pyomyo and cleaned up
